avoid using comparison endpoint for table
